requirements.txt
python-telegram-bot>=21.0
openai>=1.12.0
supabase>=2.3.0
python-dotenv>=1.0.0
Pillow>=10.4.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dateutil>=2.8.0
pytz==2024.1
APScheduler==3.10.4
nest-asyncio>=1.6.0
g4f>=0.3.9
stripe>=7.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
